+++ cli/tec.h
@@ -1,5 +1,5 @@
-#ifndef LIBTEC_TEC_H
-#define LIBTEC_TEC_H
+#ifndef TEC_TEC_H
+#define TEC_TEC_H
 
 #include <stdio.h>
 #include <getopt.h>